refactor: split release executor helpers
