+++ src/core/release/executor.rs
@@ -19,8 +19,6 @@ use crate::core::error::{Error, Result};
 use crate::core::extension::{self, ExtensionManifest};
 use crate::core::release::{changelog as release_changelog, version};
 
-use std::fmt::Write;
-
 use super::types::{ReleaseArtifact, ReleaseState, ReleaseStepResult, ReleaseStepStatus};
 use super::utils::{extract_latest_notes, parse_release_artifacts};
 
@@ -29,8 +27,11 @@ pub(crate) mod changelog;
 mod github_release;
 pub(crate) mod package_preflight;
 pub(crate) mod prepare;
+mod publish;
 mod version_targets;
 
+pub(crate) use github_release::run_github_release;
+pub(crate) use publish::{publish_response_output, run_publish};
 use version_targets::{collect_head_version_mismatches, collect_version_target_mismatches};
 
 /// Build a successful step result with optional data and hints.
@@ -73,7 +74,7 @@ fn step_failed(
 }
 
 /// Build a skipped step result carrying an explanatory warning.
-fn step_skipped(
+pub(crate) fn step_skipped(
     id: &str,
     step_type: &str,
     data: Option<serde_json::Value>,
@@ -543,313 +544,6 @@ pub(crate) fn run_package(
     Ok(step_success("package", "package", Some(data), Vec::new()))
 }
 
-/// Invoke the `release.publish` action on the named extension.
-pub(crate) fn run_publish(
-    extensions: &[ExtensionManifest],
-    state: &ReleaseState,
-    component_id: &str,
-    component_local_path: &str,
-    target: &str,
-) -> Result<ReleaseStepResult> {
-    let extension = extensions.iter().find(|m| m.id == target).ok_or_else(|| {
-        Error::validation_invalid_argument(
-            "release.publish",
-            format!("No extension '{}' found for publish target", target),
-            None,
-            Some(vec![format!(
-                "Add extension to component config: \"extensions\": {{ \"{}\": {{}} }}",
-                target
-            )]),
-        )
-    })?;
-
-    let action_id = "release.publish";
-    let has_action = extension.actions.iter().any(|a| a.id == action_id);
-    if !has_action {
-        return Err(Error::validation_invalid_argument(
-            "release.publish",
-            format!(
-                "Extension '{}' does not provide action '{}'",
-                target, action_id
-            ),
-            None,
-            None,
-        ));
-    }
-
-    let payload = build_release_payload(state, component_id, component_local_path, None);
-    let response = extension::execute_action(&extension.id, action_id, None, None, Some(&payload))?;
-    let extension_data = serde_json::to_value(&response).map_err(|e| {
-        Error::internal_json(e.to_string(), Some("extension action output".to_string()))
-    })?;
-
-    let step_id = format!("publish.{}", target);
-    let data = serde_json::json!({
-        "target": target,
-        "extension": extension.id,
-        "action": action_id,
-        "response": extension_data,
-    });
-
-    Ok(publish_step_result(
-        &step_id,
-        target,
-        &extension.id,
-        Some(data),
-        &response,
-        state.version.as_deref(),
-    ))
-}
-
-fn publish_step_result(
-    step_id: &str,
-    target: &str,
-    extension_id: &str,
-    data: Option<serde_json::Value>,
-    response: &serde_json::Value,
-    expected_version: Option<&str>,
-) -> ReleaseStepResult {
-    if let Some(reason) = extension_blocking_publish_reason(response) {
-        return step_failed(
-            step_id,
-            step_id,
-            data,
-            Some(format!(
-                "Publish to {} via {} was not completed: {}",
-                target, extension_id, reason
-            )),
-            Vec::new(),
-        );
-    }
-
-    if let Some(reason) = extension_skip_reason(response) {
-        return step_skipped(
-            step_id,
-            step_id,
-            data,
-            format!(
-                "Skipped publish to {} via {}: {}",
-                target, extension_id, reason
-            ),
-        );
-    }
-
-    if response
-        .get("success")
-        .and_then(|v| v.as_bool())
-        .unwrap_or(true)
-    {
-        if let Some(failure) = verify_publish_response(target, response, expected_version) {
-            return step_failed(step_id, step_id, data, Some(failure), Vec::new());
-        }
-
-        return step_success(step_id, step_id, data, Vec::new());
-    }
-
-    step_failed(
-        step_id,
-        step_id,
-        data,
-        Some(publish_failure_message(target, response)),
-        Vec::new(),
-    )
-}
-
-fn extension_blocking_publish_reason(response: &serde_json::Value) -> Option<String> {
-    let status = response.get("status").and_then(|v| v.as_str())?;
-    if !matches!(status, "missing_secret" | "auth_required") {
-        return None;
-    }
-
-    publish_response_reason(response).or_else(|| Some(status.replace('_', " ")))
-}
-
-fn extension_skip_reason(response: &serde_json::Value) -> Option<String> {
-    let status = response.get("status").and_then(|v| v.as_str())?;
-    if status != "skipped" {
-        return None;
-    }
-
-    publish_response_reason(response).or_else(|| Some(status.replace('_', " ")))
-}
-
-fn publish_response_reason(response: &serde_json::Value) -> Option<String> {
-    response
-        .get("reason")
-        .or_else(|| response.get("message"))
-        .and_then(|v| v.as_str())
-        .map(str::trim)
-        .filter(|s| !s.is_empty())
-        .map(str::to_string)
-        .or_else(|| {
-            let output = publish_response_output(response);
-            let detail = output.trim();
-            (!detail.is_empty()).then(|| detail.to_string())
-        })
-}
-
-fn verify_publish_response(
-    target: &str,
-    response: &serde_json::Value,
-    expected_version: Option<&str>,
-) -> Option<String> {
-    let verification = publish_registry_verification(response, expected_version)?;
-    let client = match reqwest::blocking::Client::builder()
-        .timeout(std::time::Duration::from_secs(30))
-        .build()
-    {
-        Ok(client) => client,
-        Err(err) => {
-            return Some(format!(
-                "Registry verification setup failed for {}: {}",
-                target, err
-            ));
-        }
-    };
-
-    let http_response = match client.get(&verification.url).send() {
-        Ok(response) => response,
-        Err(err) => {
-            return Some(format!(
-                "Registry verification failed for {} {}: {}",
-                verification.package_name, verification.version, err
-            ));
-        }
-    };
-
-    let status = http_response.status();
-    if !status.is_success() {
-        return Some(format!(
-            "Registry verification failed for {} {}: registry returned {}",
-            verification.package_name, verification.version, status
-        ));
-    }
-
-    let body = match http_response.text() {
-        Ok(body) => body,
-        Err(err) => {
-            return Some(format!(
-                "Registry verification failed for {} {}: could not read response: {}",
-                verification.package_name, verification.version, err
-            ));
-        }
-    };
-
-    if let Ok(json) = serde_json::from_str::<serde_json::Value>(&body) {
-        if let Some(actual_version) = json.get("version").and_then(|value| value.as_str()) {
-            if actual_version != verification.version {
-                return Some(format!(
-                    "Registry verification failed for {}: expected version {}, registry returned {}",
-                    verification.package_name, verification.version, actual_version
-                ));
-            }
-        }
-    }
-
-    None
-}
-
-#[derive(Debug)]
-struct PublishRegistryVerification {
-    package_name: String,
-    version: String,
-    url: String,
-}
-
-fn publish_registry_verification(
-    response: &serde_json::Value,
-    expected_version: Option<&str>,
-) -> Option<PublishRegistryVerification> {
-    let source = response
-        .get("registry_verification")
-        .or_else(|| response.get("registryVerification"))
-        .unwrap_or(response);
-
-    let package_name = string_field(source, &["package_name", "packageName", "package", "name"])?;
-    let version = string_field(
-        source,
-        &["version", "published_version", "publishedVersion"],
-    )
-    .or(expected_version)
-    .map(str::to_string)?;
-    let url = string_field(source, &["version_url", "versionUrl", "url"])
-        .map(str::to_string)
-        .or_else(|| {
-            let registry_url = string_field(source, &["registry_url", "registryUrl"])?;
-            Some(package_registry_version_url(
-                registry_url,
-                package_name,
-                &version,
-            ))
-        })?;
-
-    Some(PublishRegistryVerification {
-        package_name: package_name.to_string(),
-        version,
-        url,
-    })
-}
-
-fn string_field<'a>(value: &'a serde_json::Value, names: &[&str]) -> Option<&'a str> {
-    names
-        .iter()
-        .find_map(|name| value.get(*name).and_then(|value| value.as_str()))
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-}
-
-fn package_registry_version_url(registry_url: &str, package_name: &str, version: &str) -> String {
-    format!(
-        "{}/{}/{}",
-        registry_url.trim_end_matches('/'),
-        percent_encode_path_segment(package_name),
-        percent_encode_path_segment(version)
-    )
-}
-
-fn percent_encode_path_segment(value: &str) -> String {
-    let mut encoded = String::new();
-    for byte in value.bytes() {
-        match byte {
-            b'A'..=b'Z' | b'a'..=b'z' | b'0'..=b'9' | b'-' | b'.' | b'_' | b'~' => {
-                encoded.push(byte as char)
-            }
-            _ => {
-                let _ = write!(&mut encoded, "%{:02X}", byte);
-            }
-        }
-    }
-    encoded
-}
-
-fn publish_response_output(response: &serde_json::Value) -> String {
-    let stdout = response
-        .get("stdout")
-        .and_then(|v| v.as_str())
-        .unwrap_or("");
-    let stderr = response
-        .get("stderr")
-        .and_then(|v| v.as_str())
-        .unwrap_or("");
-    format!("{}\n{}", stdout, stderr)
-}
-
-fn publish_failure_message(target: &str, response: &serde_json::Value) -> String {
-    let exit_code = response
-        .get("exit_code")
-        .or_else(|| response.get("exitCode"))
-        .and_then(|v| v.as_i64());
-    let output = publish_response_output(response);
-    let detail = output.trim();
-
-    match (exit_code, detail.is_empty()) {
-        (Some(code), false) => format!("Publish to {} failed (exit {}): {}", target, code, detail),
-        (Some(code), true) => format!("Publish to {} failed (exit {})", target, code),
-        (None, false) => format!("Publish to {} failed: {}", target, detail),
-        (None, true) => format!("Publish to {} failed", target),
-    }
-}
-
 /// Delete the packaging staging dir (`target/distrib`). Skipped when the
 /// caller chose `--deploy` so the deploy step can still find the artifact.
 pub(crate) fn run_cleanup(component: &Component) -> Result<ReleaseStepResult> {
@@ -928,271 +622,6 @@ pub(crate) fn run_post_release(
         "post_release",
         "post_release",
         Some(data),
-        Vec::new(),
-    ))
-}
-
-/// Create a GitHub Release for the just-pushed tag. Fails soft in every
-/// plausible failure mode (no `gh` binary, not authenticated, release already
-/// exists, `gh release create` errors) — the tag is already pushed by the
-/// time this runs and we don't want to mark an otherwise-successful release
-/// as failed.
-pub(crate) fn run_github_release(
-    component: &Component,
-    state: &ReleaseState,
-) -> Result<ReleaseStepResult> {
-    let tag = state.tag.clone().ok_or_else(|| {
-        Error::internal_unexpected(
-            "github.release: tag state not set (git.tag must run first)".to_string(),
-        )
-    })?;
-    let notes = state.notes.clone().unwrap_or_default();
-
-    let local_path = &component.local_path;
-
-    let remote_url = component
-        .remote_url
-        .clone()
-        .or_else(|| {
-            crate::core::deploy::release_download::detect_remote_url(std::path::Path::new(
-                local_path,
-            ))
-        })
-        .ok_or_else(|| {
-            Error::internal_unexpected(
-                "github.release: no remote_url configured and git remote get-url origin failed"
-                    .to_string(),
-            )
-        })?;
-
-    let github =
-        crate::core::deploy::release_download::parse_github_url(&remote_url).ok_or_else(|| {
-            Error::validation_invalid_argument(
-                "github.release",
-                format!("Remote URL '{}' is not a GitHub URL", remote_url),
-                None,
-                Some(vec![
-                    "Only github.com remotes are supported for automatic GitHub Releases"
-                        .to_string(),
-                    "Use --no-github-release to skip this step".to_string(),
-                ]),
-            )
-        })?;
-
-    if !github_release::gh_is_available() {
-        let fallback = github_release::fallback_gh_command(&tag);
-        log_status!(
-            "release",
-            "⚠ `gh` CLI not found on PATH — skipping GitHub Release creation"
-        );
-        log_status!("release", "Manual fallback: {}", fallback);
-        return Ok(github_release::skipped_result(
-            &tag,
-            &github,
-            "gh-not-available",
-            Some(fallback),
-        ));
-    }
-
-    if !github_release::gh_is_authenticated() {
-        let fallback = github_release::fallback_gh_command(&tag);
-        log_status!(
-            "release",
-            "⚠ `gh` is not authenticated — skipping GitHub Release creation"
-        );
-        log_status!(
-            "release",
-            "Authenticate with `gh auth login`, then manual fallback: {}",
-            fallback
-        );
-        return Ok(github_release::skipped_result(
-            &tag,
-            &github,
-            "gh-not-authenticated",
-            Some(fallback),
-        ));
-    }
-
-    // Collect artifact paths from state. Populated by release.package
-    // (or any other extension action that emits artifact metadata into
-    // ReleaseState::artifacts). Passing these to `gh release create` or
-    // `gh release upload --clobber` attaches them to the Release in a
-    // single API call — keeping the github.release step responsible for
-    // the full Release lifecycle (entry + assets) instead of requiring a
-    // separate publish.<target> step.
-    let artifact_paths: Vec<String> = state
-        .artifacts
-        .iter()
-        .filter(|artifact| {
-            std::fs::metadata(&artifact.path)
-                .map(|metadata| metadata.is_file())
-                .unwrap_or(false)
-        })
-        .map(|artifact| artifact.path.clone())
-        .collect();
-    let has_artifacts = !artifact_paths.is_empty();
-
-    let repo_flag = format!("{}/{}", github.owner, github.repo);
-    if github_release::gh_release_exists(&tag, &repo_flag) {
-        // Release entry already exists (idempotent retry, or release
-        // created out of band). When the release has no artifacts to
-        // attach, skip — there is nothing to update. When artifacts are
-        // present, upload them with --clobber so retries keep the latest
-        // build attached without duplicating the GitHub Release entry.
-        if !has_artifacts {
-            log_status!(
-                "release",
-                "GitHub Release {} already exists for {} — skipping (idempotent)",
-                tag,
-                repo_flag
-            );
-            return Ok(github_release::skipped_result(
-                &tag,
-                &github,
-                "release-already-exists",
-                None,
-            ));
-        }
-
-        log_status!(
-            "release",
-            "GitHub Release {} already exists for {} — uploading {} artifact(s) with --clobber",
-            tag,
-            repo_flag,
-            artifact_paths.len()
-        );
-
-        let mut upload_args: Vec<&str> = vec!["release", "upload", &tag];
-        for path in &artifact_paths {
-            upload_args.push(path);
-        }
-        upload_args.extend_from_slice(&["--clobber", "-R", &repo_flag]);
-
-        let upload_output = std::process::Command::new("gh")
-            .args(&upload_args)
-            .output()
-            .map_err(|e| {
-                Error::internal_io(
-                    format!("Failed to invoke gh: {}", e),
-                    Some("gh release upload".to_string()),
-                )
-            })?;
-
-        if !upload_output.status.success() {
-            let stderr = String::from_utf8_lossy(&upload_output.stderr).to_string();
-            let stdout = String::from_utf8_lossy(&upload_output.stdout).to_string();
-            log_status!("release", "⚠ `gh release upload` failed: {}", stderr.trim());
-            return Ok(github_release::upload_failed_result(
-                &tag,
-                &github,
-                stdout,
-                stderr,
-                artifact_paths.len(),
-            ));
-        }
-
-        return Ok(github_release::upload_success_result(
-            &tag,
-            &github,
-            artifact_paths.len(),
-        ));
-    }
-
-    let notes_body = if notes.trim().is_empty() {
-        format!("Release {}", tag)
-    } else {
-        notes
-    };
-
-    let tmp_dir = crate::core::engine::temp::runtime_temp_dir("github-release")?;
-    let notes_path = tmp_dir.join(format!(
-        "notes-{}.md",
-        github_release::sanitize_tag_for_filename(&tag)
-    ));
-    std::fs::write(&notes_path, &notes_body).map_err(|e| {
-        Error::internal_io(
-            format!("Failed to write release notes file: {}", e),
-            Some(notes_path.display().to_string()),
-        )
-    })?;
-
-    let notes_path_str = notes_path.to_str().ok_or_else(|| {
-        Error::internal_unexpected("github.release: notes-file path is not valid UTF-8".to_string())
-    })?;
-
-    log_status!(
-        "release",
-        "Creating GitHub Release {} on {} with {} artifact(s)...",
-        tag,
-        repo_flag,
-        artifact_paths.len()
-    );
-
-    // Build args dynamically so we can append artifact paths as positional
-    // arguments — `gh release create <tag> [files...]` attaches each file
-    // as a Release asset in the same API call.
-    let mut create_args: Vec<&str> = vec![
-        "release",
-        "create",
-        &tag,
-        "--title",
-        &tag,
-        "--notes-file",
-        notes_path_str,
-        "-R",
-        &repo_flag,
-    ];
-    for path in &artifact_paths {
-        create_args.push(path);
-    }
-
-    let output = std::process::Command::new("gh")
-        .args(&create_args)
-        .output()
-        .map_err(|e| {
-            Error::internal_io(
-                format!("Failed to invoke gh: {}", e),
-                Some("gh release create".to_string()),
-            )
-        })?;
-
-    if !output.status.success() {
-        let stderr = String::from_utf8_lossy(&output.stderr).to_string();
-        let stdout = String::from_utf8_lossy(&output.stdout).to_string();
-        let fallback = github_release::fallback_gh_command(&tag);
-        log_status!("release", "⚠ `gh release create` failed: {}", stderr.trim());
-        log_status!("release", "Manual fallback: {}", fallback);
-        return Ok(step_success(
-            "github.release",
-            "github.release",
-            Some(serde_json::json!({
-                "skipped": true,
-                "reason": "gh-command-failed",
-                "tag": tag,
-                "owner": github.owner,
-                "repo": github.repo,
-                "stdout": stdout,
-                "stderr": stderr,
-                "fallback_command": fallback,
-            })),
-            Vec::new(),
-        ));
-    }
-
-    let url = String::from_utf8_lossy(&output.stdout).trim().to_string();
-    log_status!("release", "Created GitHub Release: {}", url);
-
-    Ok(step_success(
-        "github.release",
-        "github.release",
-        Some(serde_json::json!({
-            "action": "github.release",
-            "tag": tag,
-            "owner": github.owner,
-            "repo": github.repo,
-            "url": url,
-            "artifact_count": artifact_paths.len(),
-        })),
         Vec::new(),
     ))
 }
@@ -1322,17 +751,10 @@ fn store_artifacts_from_output(
 
 #[cfg(test)]
 mod tests {
-    use super::{github_release, publish_step_result, run_git_push, store_artifacts_from_output};
+    use super::{github_release, run_git_push, store_artifacts_from_output};
     use crate::core::component::Component;
     use crate::core::release::ReleaseStepStatus;
-    use std::io::{Read, Write};
-    use std::net::TcpListener;
     use std::process::Command;
-
-    struct RegistryServer {
-        registry_url: String,
-        handle: std::thread::JoinHandle<String>,
-    }
 
     fn git(path: &std::path::Path, args: &[&str]) {
         let output = Command::new("git")
@@ -1346,36 +768,6 @@ mod tests {
             args,
             String::from_utf8_lossy(&output.stderr)
         );
-    }
-
-    fn registry_server(status: u16, body: &'static str) -> RegistryServer {
-        let listener = TcpListener::bind("127.0.0.1:0").expect("bind registry server");
-        let port = listener.local_addr().expect("registry server addr").port();
-        let registry_url = format!("http://127.0.0.1:{}", port);
-        let handle = std::thread::spawn(move || {
-            let (mut stream, _) = listener.accept().expect("accept registry request");
-            let mut buffer = [0; 1024];
-            let bytes_read = stream.read(&mut buffer).expect("read registry request");
-            let request = String::from_utf8_lossy(&buffer[..bytes_read]).to_string();
-            let request_line = request.lines().next().unwrap_or_default().to_string();
-            let reason = if status == 200 { "OK" } else { "Not Found" };
-            let response = format!(
-                "HTTP/1.1 {} {}\r\ncontent-type: application/json\r\ncontent-length: {}\r\nconnection: close\r\n\r\n{}",
-                status,
-                reason,
-                body.len(),
-                body
-            );
-            stream
-                .write_all(response.as_bytes())
-                .expect("write registry response");
-            request_line
-        });
-
-        RegistryServer {
-            registry_url,
-            handle,
-        }
     }
 
     #[test]
@@ -1488,127 +880,6 @@ mod tests {
         assert_eq!(result.status, ReleaseStepStatus::Success);
         git(remote.path(), &["show-ref", "--verify", "refs/heads/main"]);
         git(remote.path(), &["show-ref", "--verify", "refs/tags/v1.0.0"]);
-    }
-
-    #[test]
-    fn publish_step_fails_when_extension_reports_missing_secret() {
-        let response = serde_json::json!({
-            "success": false,
-            "status": "missing_secret",
-            "reason": "registry token is not configured",
-            "stdout": "",
-            "stderr": "",
-        });
-        let data = serde_json::json!({ "response": response.clone() });
-
-        let result = publish_step_result(
-            "publish.registry",
-            "registry",
-            "registry",
-            Some(data),
-            &response,
-            None,
-        );
-
-        assert_eq!(result.status, ReleaseStepStatus::Failed);
-        assert!(result
-            .error
-            .unwrap()
-            .contains("registry token is not configured"));
-        assert!(result.warnings.is_empty());
-    }
-
-    #[test]
-    fn publish_step_fails_when_extension_reports_auth_required() {
-        let response = serde_json::json!({
-            "success": false,
-            "status": "auth_required",
-            "message": "run the extension login command",
-        });
-
-        let result = publish_step_result(
-            "publish.registry",
-            "registry",
-            "registry",
-            None,
-            &response,
-            None,
-        );
-
-        assert_eq!(result.status, ReleaseStepStatus::Failed);
-        assert!(result
-            .error
-            .unwrap()
-            .contains("run the extension login command"));
-    }
-
-    #[test]
-    fn publish_step_fails_when_extension_error_has_no_skip_status() {
-        let response = serde_json::json!({
-            "success": false,
-            "exitCode": 1,
-            "stderr": "error: failed to upload package: 500 server error",
-        });
-
-        let result = publish_step_result(
-            "publish.registry",
-            "registry",
-            "registry",
-            None,
-            &response,
-            None,
-        );
-
-        assert_eq!(result.status, ReleaseStepStatus::Failed);
-        assert!(result.error.unwrap().contains("500 server error"));
-    }
-
-    #[test]
-    fn publish_step_fails_when_registry_version_is_missing() {
-        let server = registry_server(404, r#"{"error":"not found"}"#);
-        let response = serde_json::json!({
-            "success": true,
-            "registry_verification": {
-                "registry_url": server.registry_url,
-                "package_name": "@extrachill/components",
-                "version": "0.5.2"
-            }
-        });
-
-        let result =
-            publish_step_result("publish.npm", "npm", "npm", None, &response, Some("0.5.2"));
-
-        assert_eq!(result.status, ReleaseStepStatus::Failed);
-        assert!(result
-            .error
-            .unwrap()
-            .contains("registry returned 404 Not Found"));
-        assert_eq!(
-            server.handle.join().expect("server request"),
-            "GET /%40extrachill%2Fcomponents/0.5.2 HTTP/1.1"
-        );
-    }
-
-    #[test]
-    fn publish_step_succeeds_when_registry_version_is_present() {
-        let server = registry_server(200, r#"{"version":"0.5.2"}"#);
-        let response = serde_json::json!({
-            "success": true,
-            "registry_verification": {
-                "registry_url": server.registry_url,
-                "package_name": "@extrachill/components",
-                "version": "0.5.2"
-            }
-        });
-
-        let result =
-            publish_step_result("publish.npm", "npm", "npm", None, &response, Some("0.5.2"));
-
-        assert_eq!(result.status, ReleaseStepStatus::Success);
-        assert_eq!(
-            server.handle.join().expect("server request"),
-            "GET /%40extrachill%2Fcomponents/0.5.2 HTTP/1.1"
-        );
     }
 
     #[test]
